Create main function

+++ Observer/ConcreteObserver.cpp
@@ -2,6 +2,6 @@
 #include <iostream>
 void onNotify(const Entity& t_entity, Event t_event)
 {
-	std::cout << "Observed Event with ID" << t_event.id << std::endl;
+	std::cout << "Observed Event with ID" << (int)t_event.id << std::endl;
 	std::cout << "Entity Name: " << t_entity.name << std::endl;
 }
